requirements.txt
pyyaml
telepot
